refactor(rust/sedona-raster-gdal): extract reusable tiling core from RS_Tile
Pure refactor, no behavior change. Move RS_Tile's tile-at-a-time tiling
algorithm (tile_grid_dims, TileWindow, resolve_band_indices, append_tile
and its private pixel helpers, TileParams) into a new crate-internal
sibling module `tiling.rs`, exposed as `pub(crate)` so a future streaming
tile-generating operator can drive it one tile at a time. `rs_tile.rs`
keeps only the RS_Tile UDF surface (registration, overload ladder, per-row
argument parsing, List<Struct<x, y, tile>> assembly); explode_raster_tiles
becomes a thin driver over the tiling core.

Tiles, ordering, nodata/padding, and errors are unchanged. The pure
tiling-algorithm tests moved into `tiling.rs`'s own test module (moves,
not deletions); the UDF-surface tests stay in `rs_tile.rs`.
