+++ rust/sedona-raster-gdal/src/lib.rs
@@ -45,6 +45,7 @@ mod rs_reproject_match;
 mod rs_resample;
 mod rs_tile;
 mod rs_zonal_stats;
+mod tiling;
 mod utils;
 
 // Re-export main dataset conversion functions
